requirements.txt
grpcio==1.35.0
grpcio-tools==1.35.0
protobuf==3.14.0
snappi==0.3.14
snappi[ixnetwork]==0.1.39